break main.py into more modular files, add directory structure for better organization
