+++ app/main.py
@@ -1,0 +1,4 @@
+from app.api import api_handlers
+
+if "__name__" == "__main__":
+    app = api_handlers.app
